monitoring_backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
python-decouple>=3.6
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
whitenoise>=6.0.0
